Rename test. Add utility method that creates HttpHealthCheckedEndpointGroup instance
