+++ core/src/main/java/com/linecorp/armeria/client/routing/endpoint/healthcheck/HttpHealthCheckedEndpointGroup.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2016 LINE Corporation
+ *
+ * LINE Corporation licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
 package com.linecorp.armeria.client.routing.endpoint.healthcheck;
 
 import static java.util.Objects.requireNonNull;
@@ -11,46 +26,69 @@ import com.linecorp.armeria.client.http.HttpClient;
 import com.linecorp.armeria.client.routing.EndpointGroup;
 import com.linecorp.armeria.common.http.HttpStatus;
 
-public class HttpHealthCheckedEndpointGroup extends HealthCheckedEndpointGroup {
+/**
+ * HTTP implementation of {@link HealthCheckedEndpointGroup}.
+ */
+public final class HttpHealthCheckedEndpointGroup extends HealthCheckedEndpointGroup {
     private final String healthCheckPath;
 
     /**
-     * Creates a new instance.
+     * Creates a new {@link HttpHealthCheckedEndpointGroup} instance.
      */
-    public HttpHealthCheckedEndpointGroup(ClientFactory clientFactory,
-                                          EndpointGroup delegate,
-                                          long healthCheckRetryDelayMills,
-                                          String healthCheckPath) {
+    private HttpHealthCheckedEndpointGroup(ClientFactory clientFactory,
+                                           EndpointGroup delegate,
+                                           String healthCheckPath,
+                                           long healthCheckRetryDelayMills) {
         super(clientFactory, delegate, healthCheckRetryDelayMills);
         this.healthCheckPath = requireNonNull(healthCheckPath, "healthCheckPath");
+        init();
     }
 
     /**
-     * Creates a new instance.
+     * Creates a new {@link HttpHealthCheckedEndpointGroup} instance.
      */
-    public HttpHealthCheckedEndpointGroup(EndpointGroup delegate,
-                                          String healthCheckPath,
-                                          long healthCheckRetryDelayMills) {
-        this(ClientFactory.DEFAULT, delegate, healthCheckRetryDelayMills, healthCheckPath);
+    public static HttpHealthCheckedEndpointGroup of(EndpointGroup delegate,
+                                                    String healthCheckPath) {
+        return of(delegate, healthCheckPath, DEFAULT_HEALTHCHECK_RETRY_DELAY_MILLS);
     }
 
+    /**
+     * Creates a new {@link HttpHealthCheckedEndpointGroup} instance.
+     */
+    public static HttpHealthCheckedEndpointGroup of(EndpointGroup delegate,
+                                                    String healthCheckPath,
+                                                    long healthCheckRetryDelayMills) {
+        return of(ClientFactory.DEFAULT, delegate, healthCheckPath, healthCheckRetryDelayMills);
+    }
+
+    /**
+     * Creates a new {@link HttpHealthCheckedEndpointGroup} instance.
+     */
+    public static HttpHealthCheckedEndpointGroup of(ClientFactory clientFactory,
+                                                    EndpointGroup delegate,
+                                                    String healthCheckPath,
+                                                    long healthCheckRetryDelayMills) {
+        return new HttpHealthCheckedEndpointGroup(clientFactory,
+                                                  delegate,
+                                                  healthCheckPath,
+                                                  healthCheckRetryDelayMills);
+    }
     @Override
     EndpointHealthChecker createEndpointHealthChecker(Endpoint endpoint) {
         return new HttpEndpointHealthChecker(clientFactory, endpoint, healthCheckPath);
     }
 
     private static final class HttpEndpointHealthChecker implements EndpointHealthChecker {
-
-        private final String healthCheckPath;
         private final HttpClient httpClient;
+        private final String healthCheckPath;
 
-        HttpEndpointHealthChecker(ClientFactory clientFactory,
-                                  Endpoint endpoint,
-                                  String healthCheckPath) {
+        private HttpEndpointHealthChecker(ClientFactory clientFactory,
+                                          Endpoint endpoint,
+                                          String healthCheckPath) {
+            httpClient = Clients.newClient(clientFactory,
+                                           "none+http://" + endpoint.authority(),
+                                           HttpClient.class);
             this.healthCheckPath = healthCheckPath;
-            this.httpClient = Clients.newClient(clientFactory,
-                                                "none+http://" + endpoint.authority(),
-                                                HttpClient.class);
         }
 
         @Override
